// src/meta/generated/filters.ts
export interface BaseFilterConfiguration<T extends string> {
  type: T;
  display?: string;
}

export interface DateIntervalFilterConfiguration extends BaseFilterConfiguration<'dateInterval'> {}

export interface TextFilterConfiguration extends BaseFilterConfiguration<'text'> {
  trim?: boolean;
  lowercase?: boolean;
  minLen?: number;
  maxLen?: number;
}

export interface NumberFilterConfiguration extends BaseFilterConfiguration<'number'> {
  min?: number;
  max?: number;
}

export interface RangeFilterConfiguration extends BaseFilterConfiguration<'range'> {
  min?: number;
  max?: number;
}

export interface BooleanFilterConfiguration extends BaseFilterConfiguration<'boolean'> {
  default?: boolean;
}

export interface SelectFilterConfiguration<T extends string> extends BaseFilterConfiguration<'select' | 'dropdown'> {
  options: {
    value: T;
    display?: string;
  }[];
  multiple?: boolean;
}

export interface SpecificFilterConfiguration extends BaseFilterConfiguration<'specific_dropdown'> {
  data_url: string;
}

export type SelectFilterOptionsValues<T> = T extends {
  options: { value: infer S }[];
}
  ? S
  : never;

export type FilterConfiguration =
  | DateIntervalFilterConfiguration
  | TextFilterConfiguration
  | NumberFilterConfiguration
  | RangeFilterConfiguration
  | BooleanFilterConfiguration
  | SelectFilterConfiguration<string>
  | SpecificFilterConfiguration;

export type FilterConfigurationFor<T extends FilterConfiguration['type']> = Omit<
  FilterConfiguration & { type: T },
  'type'
>;
/// region: type-map-end

export type FilterDefinition = {
  name: string;
  properties: Record<string, FilterConfiguration>;
};